requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
branca==0.4.2
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
Flask==1.1.2
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-Scss==0.5
folium==0.12.1
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mysqlclient @ file:///D:/Downloads/mysqlclient-1.4.6-cp39-cp39-win_amd64.whl
numpy==1.20.1
packaging==20.9
pandas==1.2.2
Pillow==8.1.0
plotly==4.14.3
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pyScss==1.3.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
retrying==1.3.3
scipy==1.6.0
seaborn==0.11.1
six==1.15.0
toml==0.10.2
urllib3==1.26.3
Werkzeug==1.0.1
